% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTraj.R
\name{getTraj}
\alias{getTraj}
\title{Get backward trajectories}
\usage{
getTraj(sigmaU, sigmaV, sigmaW, ustar, ubar, flux, L, u, v, w, nTraj,
  outSteps, height_sens, MaxFetch = 3000)
}
\description{
Get backward trajectories
}
