% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_latlon.R
\name{grid_to_latlon}
\alias{grid_to_latlon}
\title{Makes raster grid for aggregating trajecotries}
\usage{
grid_to_latlon(rast_grid, deg = TRUE)
}
\arguments{
\item{site_coords}{vector of lon, lat}

\item{grid_res}{resolution of grid cells}

\item{max_dist}{maximum distance from center point in cardinal directions}
}
\description{
\code{make_grid}
}
