% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{make_raster}
\alias{make_raster}
\title{Makes raster grid for aggregating trajecotries}
\usage{
make_raster(site_coords, grid_res, max_dist)
}
\arguments{
\item{site_coords}{vector of lon, lat}

\item{grid_res}{resolution of grid cells}

\item{max_dist}{maximum distance from center point in cardinal directions}
}
\description{
\code{make_grid}
}
