% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonicQA.R
\name{sonicQA}
\alias{sonicQA}
\title{Remove Spikes from Sonic Data}
\usage{
sonicQA(spod, nodes)
}
\arguments{
\item{spod}{data.frame with SPod data}

\item{nodes}{vector of letters with spod nodes}
}
\description{
Used to remove sharp jumps in sonic data
}
