% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrotate.R
\name{unrotate}
\alias{unrotate}
\title{Rotate back to orginal coordinate system}
\usage{
unrotate(u_f, v_f, w_f, angles)
}
\arguments{
\item{u_f}{along wind coordinate}

\item{v_f}{cross wind coordinate}

\item{w_f}{vertical coordianate}

\item{angles}{3 rotation angles used to streamline coordinates}
}
\description{
Rotate back to orginal coordinate system
}
